♻️  router research_loop.py 추출 — P0-K guard + hook + forum status (P0-P step 9)
변경 이유
- ``_run_research_loop_hook`` + P0-K command-only 가드 +
  ``persist_research_forum_status`` + ``make_default_research_loop``
  factory + member-bots forum status formatter 가 ``_legacy`` 안에서
  ~600 줄을 차지하면서 forum publish / member-bot 모드 / research
  pack persistence 까지 한 묶음으로 entangled.

주요 변경 사항
- 새 파일 ``research_loop.py`` (약 700 줄):
  - ``_research_loop_blocked_by_command_only`` (P0-K hard rule)
  - ``_run_research_loop_hook`` — research loop 호출 + 결과 surface
  - ``_maybe_persist_research_pack`` — session.extra 적재
  - ``persist_research_forum_status`` — 운영자 status 라인
  - ``_format_member_bots_forum_status`` — member-bots 모드 디스플레이
  - ``make_default_research_loop`` — production hook factory
- ``_legacy.py`` 의 해당 6 함수 (486 줄) 제거 → re-export.
- ``persist_research_artifacts`` import + ``_coerce_research_loop_report``
  lazy shim (step 10 에서 reporting 추출 시 정리 예정).
- 테스트 mock 경로 ``_legacy.persist_research_artifacts`` →
  ``research_loop.persist_research_artifacts`` 갱신 3 곳.

비고
- 4672 PASS + 1 skipped 회귀 유지.
- P0-K hard rule (3 call site) 무회귀 — guard 함수 자체와 caller 모두
  같은 모듈 내에서 유지되어 가드 일관성 보장.
